Fix 41/NOT_CONTROLLER for non-Kraft clusters
For non-Kraft clusters requests to create and delete clusters must be
sent to the controller broker. Non-controller brokers are answering
such requests with error 41/NOT_CONTROLLER. This change ensures that
those requests are sent to the correct controller in case we receive
error code 41 upon creation or deletion of topics.

on-behalf-of: @SAP ospo@sap.com