Do not deduplicate topic creations and deletions
As of now creating and deleting multiple topics in a row will ignore
most of the topics except the first one. This is due to deduplication
that is done based on operation alone without considering the target
topics. This change incorporates topic data into the deduplication to
avoid the aforementioned issue.

on-behalf-of: @SAP ospo@sap.com