Adjust default values for replicas and partitions
Requests to create topics will fail if you supply assignments and
partitions/replicas together. If you don't supply a value for
partitions/replicas at all, default values of 1 will be used which also
leads to topic creation failures. Here, the default values for the two
settings are changed to -1 so that topic creations do not fail due to
conflicting settings. Furthermore, this has the benefit of using the
default values set inside the Kafka cluster.

Related:
- https://github.com/tulios/kafkajs/pull/1305