Fix unhandled errors on custom connection errors
This change fixes crashes due to unhandled errors if more than one
packet with unknown correlation ID are received in succession. The
connection pool attaches an error handler via
`connection.once('error', ...)` which detaches the error handler after
being invoked once. Further emitted errors will be handled as uncaught
exceptions and exit the program. Also, the socket is not closed when
emitting this error and keeps listening for incoming messages. By
destroying the socket and providing the appropriate error we prevent
the socket from handling any more packets, thus preventing uncaught
exceptions.

Furthermore, by providing the error as parameter to the socket
destruction we stay within the normal socket lifecycle and all
clean up on socket closure is done, which was previously skipped
entirely.

on-behalf-of: @SAP ospo@sap.com
Signed-off-by: Peter Lehnhardt <peter.lehnhardt@sap.com>